fix: issue in consuming interrupt
whew, debugging journey and I had to rely on the reference implementation

The use of take() to cause a side-effect during polling
was what had confused me. I didn't know `take` was a mutating method,
so I had written the code differently. My code was causing interrupts
to fire in a loop, and eventually overrun the NES CPU's stack